Also catch RuntimeError when importing PyPDF so that PyPDF or, crucially, its transitive dependencies do not cause diffoscope to traceback at runtime and build time. (Closes: Debian:#1078944, reproducible-builds/diffoscope#389)
The root or even primary cause of this bug is not in diffoscope itself; rather,
the cause is to be found between src:python3-cryptography and src:openssl — see
bugs #1078747 and #965041 for more details.

However, diffoscope should (and easily can…) be mindful that importing external
libraries could result in exceptions other than ImportError when we have reason
to believe that they might do so.

We already do so in fsimage.py for a complicated library (guestfs), so I think
it's not at all unreasonable to catch it here, too.

Alas I can't seem to reproduce this on my local machine or in a clean sid
chroot, so I am uploading the following patch after simply having a hard stare
at the code… albeit on a warm Sunday afternoon.