Add support for UKI files. (Closes: reproducible-builds/diffoscope!146)
Diffoscope supported diffing UKI's with objdump in Pe32PlusFile, but
this doesn't work too well with 200M+ UKI's. It takes a long time to
generate the diff, ukify from systemd-ukify gives a user-friendly dump
of the sections with hashes so one can easily see what part of the UKI
is not reproducible.

Signed-off-by: Chris Lamb <lamby@debian.org>