rpm: fix compat with rpm 6
The python interface in rpm 5.99 (pre-release for 6) dropped
RPMTAG_HEADERSIGNATURES, because it is not used anymore. But we need
to work with old and new rpms, so we need that value. Let's just copy
the enum values into the code. Those values can never change (or be
reused) because they are used in countless rpms.